# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Thundergate API
  description: |
    Thundergate — a real-time execution firewall between autonomous AI agents
    and external APIs. Provides rule-based evaluation, human-in-the-loop
    review, and tamper-evident audit logging.
  version: 0.1.0
  contact:
    name: Thundergate Team
servers:
  - url: http://localhost:3001
    description: Proxy (Agent-facing API)
  - url: http://localhost:3000
    description: Dashboard (Operator-facing API)
...
paths:
...
  /queue/{queueId}/decide:
    post:
      tags: [queue]
      summary: Submit operator decision on a queue item
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, MODIFIED, REJECTED]
                notes:
                  type: string
                modifiedPayload:
                  type: object
      responses:
        '200':
          description: Decision recorded

  # ─── Dashboard API — Queue (port 3000) ──────────────────
...
  /api/queue/{id}:
    get:
      tags: [queue]
      summary: Get queue item detail with audit log context
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Queue item with audit log
        '404':
          description: Queue item not found
...
  /api/queue/{id}/decide:
    post:
      tags: [queue]
      summary: Submit decision on a claimed queue item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, MODIFIED, REJECTED]
                notes:
                  type: string
                modifiedPayload:
                  type: object
      responses:
        '200':
          description: Decision recorded
...
  /api/agents:
    get:
      tags: [agents]
      summary: List all registered agents
      responses:
        '200':
          description: Array of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      tags: [agents]
      summary: Register a new agent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Agent created with API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  apiKey:
                    type: string
                    description: Plain-text API key (shown once)
  /api/agents/{id}:
    get:
      tags: [agents]
      summary: Get agent detail with statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent detail
        '404':
          description: Agent not found
  /api/agents/{id}/toggle:
    post:
      tags: [agents]
      summary: Activate or deactivate an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive:
                  type: boolean
      responses:
        '200':
          description: Agent status updated
  /api/agents/{id}/rotate-key:
    post:
      tags: [agents]
      summary: Rotate agent API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: New API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey:
                    type: string
...
  /api/audit:
    get:
      tags: [audit]
      summary: List audit logs with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: decision
          in: query
          schema:
            type: string
            enum: [ALLOW, BLOCK, FLAG_FOR_REVIEW, MODIFY]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, riskScore, latencyMs]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
  /api/audit/{id}:
    get:
      tags: [audit]
      summary: Get audit log detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full audit log entry
  /api/audit/verify:
    post:
      tags: [audit]
      summary: Verify hash chain integrity
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dateFrom:
                  type: string
                  format: date-time
                dateTo:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  checked:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
  /api/audit/export:
    get:
      tags: [audit]
      summary: Export audit logs as CSV or JSON
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
            default: json
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: File download
...
components:
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        requestCount24h:
          type: integer
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequenceNumber:
          type: integer
        agentId:
          type: string
          format: uuid
        agentName:
          type: string
        requestMethod:
          type: string
        requestUrl:
          type: string
        riskScore:
          type: integer
        engineDecision:
          type: string
          enum: [ALLOW, BLOCK, FLAG_FOR_REVIEW, MODIFY]
        humanDecision:
          type: string
          enum: [APPROVED, MODIFIED, REJECTED, EXPIRED]
          nullable: true
        latencyMs:
          type: integer
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    AgentApiKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: Agent API key for proxy authentication
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Dashboard session cookie (NextAuth)